api/place_order/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pika==1.2.0
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
requests==2.27.1